+++ Nagisa.Shared/M2.Helpers.WinRT.cpp
@@ -324,3 +324,33 @@ string M2MakeUTF8String(String^& PlatformString)
 {
 	return M2MakeUTF8String(M2MakeUTF16String(PlatformString));
 }
+
+// The M2MakeUTF8String function finds a sub string from a source string.
+//
+// Parameters:
+//
+// SourceString
+//     The source string.
+// SubString
+//     The sub string.
+// IgnoreCase
+//     The option can determines whether to ignore case.
+//
+// Return value:
+//
+// If success, it will return true.
+bool M2FindSubString(
+	String^& SourceString, String^& SubString, bool IgnoreCase)
+{
+	return (::FindNLSStringEx(
+		nullptr,
+		(IgnoreCase ? NORM_IGNORECASE : 0) | FIND_FROMSTART,
+		SourceString->Data(),
+		SourceString->Length(),
+		SubString->Data(),
+		SubString->Length(),
+		nullptr,
+		nullptr,
+		nullptr,
+		0) >= 0);
+}